Remake of Minecraft Tickets functionality
Mc tickets part is now fully automatic and completed 100%

+++ server/src/main/java/com/scoutress/KaimuxAdminStats/servicesImpl/ApiDataExtractionServiceImpl.java
@@ -119,7 +119,7 @@ public class ApiDataExtractionServiceImpl implements ApiDataExtractionService {
       }
 
       MinecraftTicketsAnswers entity = new MinecraftTicketsAnswers();
-      entity.setMinecraftTicketId(item.getLong("player_id"));
+      entity.setMinecraftTicketId((short) item.getInt("player_id"));
       entity.setDateTime(dateTime);
 
       minecraftTicketsAnswersRepository.save(entity);
